Crate people_rococo_runtime

Modules

api

people

xcm_config

Structs

DepositBase

DepositFactor

ExistentialDeposit

FellowsBodyId

MessageQueueServiceWeight

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

Perbill

A fixed point representation of a number in the range [0, 1].

Permill

A fixed point representation of a number in the range [0, 1].

PotId

RelayOrigin

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SS58Prefix

SessionKeys

SessionLength

StakingAdminBodyId

TransactionByteFee

Relay Chain TransactionByteFee / 10.

Version

Enums

MultiAddress

A multi-format address wrapper for on-chain accounts.

OriginCaller

RuntimeCall

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

OFFSET

PERIOD

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Address

The address format for describing accounts.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

We allow Root and the StakingAdmin to execute privileged collator selection operations.

CumulusXcm

Executive

Executive: handles dispatch to the various modules.

GenesisConfigDeprecated

Identity

IdentityMigrator

MessageQueue

Migrations

Migrations to apply on runtime upgrade.

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

RootOrFellows

Privileged origin that represents Root or Fellows pluralistic body.

Session

SessionConfig

SignedBlock

A Block signed with an sp_runtime::Justification.

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

XcmpQueue