Module people_rococo_runtime::xcm_config

Structs

BaseDeliveryFee

The base fee for the message delivery fees.

FeeAssetId

The asset ID for the asset that we use to pay for message delivery fees. Just ROC.

FellowshipLocation

GovernanceLocation

LocalPlurality

MaxAssetsIntoHolding

MaxInstructions

ParentOrParentsPlurality

RelayChainOrigin

RelayLocation

RelayNetwork

RelayTreasuryLocation

RootLocation

TreasuryAccount

UniversalLocation

XcmConfig

Type Aliases

Barrier

FungibleTransactor

Means for transacting the native currency on this chain.

LocalOriginToLocation

Converts a local signed origin into an XCM location. Forms the basis for local origins sending/executing XCMs.

LocationToAccountId

Type for specifying how a Location can be converted into an AccountId. This is used when determining ownership of accounts for asset transacting and when attempting to use XCM Transact in order to determine the dispatch Origin.

PriceForParentDelivery

PriceForSiblingParachainDelivery

WaivedLocations

Locations that will not be charged fees in the executor, neither for execution nor delivery. We only waive fees for system functions, which these locations represent.

XcmOriginToTransactDispatchOrigin

This is the type we use to convert an (incoming) XCM origin into a local Origin instance, ready for dispatching a transaction with XCM’s Transact. There is an OriginKind that can bias the kind of local Origin it will become.

XcmRouter

The means for routing XCM messages which are not for local execution into the right message queues.